backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
stripe==7.8.0
supabase==2.3.4
openai==1.3.7
anthropic==0.7.8
langchain==0.0.340
langchain-openai==0.0.2
together==0.2.7